(osm2vt) First pass at `Way.toGeometry`
- Although this probably doesn't work. This commit is a checkpoint before I
put together some unit tests.
